pyproject.toml
[project]
name = "dio_bank"
version = "0.1.0"
description = ""
authors = [
    {name = "mlkrinski",email = "marcelokrinski@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask",
    "flask-sqlalchemy",
    "flask-migrate",
    "flask-jwt-extended",
    "psycopg2-binary",
    "gunicorn",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-bcrypt"
]

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = ["tests"]



[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
